Refactor backend modules
